Rearrange realm root contexts a bit
Execution stack no longer keeps the realm's root context at the bottom always.  Instead, the context is held by the RealmRecord and we require the realm to be passed into $.evaluateScript() so that it can enter the context before running the script (we will tack on a microtask queue flush at the end of the evaluation, which might change the realm).